requirements.txt
pre-commit==23.1.0
Pillow
click
shutil
progressbar2
